+++ app-backend/project/Dependencies.scala
@@ -140,8 +140,7 @@ object Dependencies {
   val doobieHikari = "org.tpolecat" %% "doobie-hikari" % Version.doobie
   val doobiePostgres = "org.tpolecat" %% "doobie-postgres" % Version.doobie
   val doobiePostgresCirce = "org.tpolecat" %% "doobie-postgres-circe" % Version.doobie
-  val doobieScalatest = "org.tpolecat" %% "doobie-scalatest" % Version.doobie
-  val doobieSpecs = "org.tpolecat" %% "doobie-specs2" % Version.doobie
+  val doobieScalatest = "org.tpolecat" %% "doobie-scalatest" % Version.doobie % "test"
   val dropbox = "com.dropbox.core" % "dropbox-core-sdk" % Version.dropbox
   val elasticacheClient = "com.amazonaws" % "elasticache-java-cluster-client" % Version.elasticacheClient
   val ficus = "com.iheart" %% "ficus" % Version.ficus